Move XR scripts into scripts/esm/xr folder
